Abstract React Query into custom hooks
