+++ src/features/cabins/CreateCabinForm.jsx
@@ -4,14 +4,12 @@ import Button from "../../ui/Button";
 import FileInput from "../../ui/FileInput";
 import Textarea from "../../ui/Textarea";
 import { useForm } from "react-hook-form";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { createEditCabin } from "../../services/apiCabins";
-import toast from "react-hot-toast";
 import FormRow from "../../ui/FormRow";
-// import PropType from 'prop-types';
+import { useCreateCabin } from "./useCreateCabin";
+import { useEditCabin } from "./useEditCabin";
 
 // eslint-disable-next-line react/prop-types
-function CreateCabinForm({ cabinToEdit = {} }) {
+function CreateCabinForm({ cabinToEdit={} }) {
   const { id: editId, ...editValues } = cabinToEdit;
   const isEditSession = Boolean(editId); 
 
@@ -20,34 +18,23 @@ function CreateCabinForm({ cabinToEdit = {} }) {
   });
 
   const { errors } = formState;
-  const queryClient = useQueryClient();
 
-  const {mutate: createCabin, isLoading: isCreating} = useMutation({
-    mutationFn: createEditCabin,
-    onSuccess: () => {
-      toast.success("New cabin successfully created");
-      queryClient.invalidateQueries({ queryKey: ["cabins"] });
-      reset();
-    },
-    onError: (err) => toast.error(err.message),
-  });
+  const { isCreating, createCabin } = useCreateCabin();
 
-  const {mutate: editCabin, isLoading: isEditing} = useMutation({
-    mutationFn: ({ newCabinData, id}) => createEditCabin(newCabinData, id),
-    onSuccess: () => {
-      toast.success("Cabin successfully edited");
-      queryClient.invalidateQueries({ queryKey: ["cabins"] });
-      reset();
-    },
-    onError: (err) => toast.error(err.message),
-  });
+  const { isEditing, editCabin } = useEditCabin();
 
   const isWorking = isCreating || isEditing;
 
   function onSubmit(data) {
     const image = typeof data.image === "string" ? data.image : data.image[0];
     if (isEditSession)
-    editCabin({ newCabinData: {...data, image }, id: editId });
+    editCabin({ newCabinData: {...data, image }, id: editId },
+      {
+        onSuccess: (data) => {
+          reset();
+        },
+      }
+  );
     else createCabin({ ...data, image: image });
     // return true;
   }
@@ -109,9 +96,5 @@ function CreateCabinForm({ cabinToEdit = {} }) {
     </Form>
   );
 }
-
-// CreateCabinForm.propType = {
-//   cabinToEdit: PropType.object.isRequired,
-// }
 
 export default CreateCabinForm;
