+++ src/features/cabins/useCreateCabin.js
@@ -1,0 +1,17 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { createEditCabin } from "../../services/apiCabins";
+import toast from "react-hot-toast";
+
+export function useCreateCabin() {
+  const queryClient = useQueryClient();
+  const {mutate: createCabin, isLoading: isCreating} = useMutation({
+    mutationFn: createEditCabin,
+    onSuccess: () => {
+      toast.success("New cabin successfully created");
+      queryClient.invalidateQueries({ queryKey: ["cabins"] });
+      reset();
+    },
+    onError: (err) => toast.error(err.message),
+  });
+  return { isCreating, createCabin };
+}
