+++ src/features/cabins/useEditCabin.js
@@ -1,0 +1,20 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { createEditCabin } from "../../services/apiCabins";
+import toast from "react-hot-toast";
+
+export function useEditCabin() {
+
+const queryClient = useQueryClient();
+
+
+  const {mutate: editCabin, isLoading: isEditing} = useMutation({
+    mutationFn: ({ newCabinData, id}) => createEditCabin(newCabinData, id),
+    onSuccess: () => {
+      toast.success("Cabin successfully edited");
+      queryClient.invalidateQueries({ queryKey: ["cabins"] });
+      reset();
+    },
+    onError: (err) => toast.error(err.message),
+  });
+  return { isEditing, editCabin };
+}
